Move FluxRuntimeObjects handler to its own file
